Align optimization APIs and course tests
