+++ src/tiny_llm/week2_kernels.py
@@ -1,0 +1,37 @@
+import mlx.core as mx
+
+
+class FastRMSNorm:
+    def __init__(self, dim: int, weight: mx.array, eps: float = 1e-5):
+        pass
+
+    def __call__(self, x: mx.array) -> mx.array:
+        pass
+
+
+class FastRoPE:
+    def __init__(
+        self,
+        dims: int,
+        seq_len: int,
+        base: int = 10000,
+        traditional: bool = False,
+    ):
+        pass
+
+    def __call__(self, x: mx.array, offset: int | list[int] | mx.array = 0) -> mx.array:
+        pass
+
+
+def swiglu(gate: mx.array, up: mx.array) -> mx.array:
+    pass
+
+
+def scaled_dot_product_attention(
+    query: mx.array,
+    key: mx.array,
+    value: mx.array,
+    scale: float,
+    mask: mx.array | str | None = None,
+) -> mx.array:
+    pass
