+++ tests_refsol/test_week_2_day_6.py
@@ -1,335 +1,106 @@
+"""Week 2 end-to-end optimized model tests."""
+
 import mlx.core as mx
-import numpy as np
 import pytest
+from .utils import *
+from .tiny_llm_base import (
+    Qwen3ModelWeek2,
+    Qwen3ModelWeek3,
+    TinyKvFullCache,
+)
 from mlx_lm import load
 
-from .tiny_llm_base import *
-from .utils import *
+# TODO: task 1 tests
 
 
-def rope_helper(stream: mx.Stream, traditional: bool, precision: mx.Dtype):
-    BATCH_SIZE = 16
-    NUM_HEADS = 8
-    HEAD_DIM = 4
-    MAX_SEQ_LEN = 14
-    SEQ_LEN = 9
-    BASE = 10000
-    with mx.stream(stream):
-        for _ in range(100):
-            user_layer = RoPE(HEAD_DIM, MAX_SEQ_LEN, BASE, traditional=traditional)
-            x = mx.random.uniform(
-                shape=(BATCH_SIZE, SEQ_LEN, NUM_HEADS, HEAD_DIM), dtype=precision
-            )
-
-            input_pos = np.random.randint(0, MAX_SEQ_LEN - SEQ_LEN, size=BATCH_SIZE)
-            input_pos_mx = mx.array(input_pos, dtype=mx.int32)
-            input_pos_user = [slice(i, i + SEQ_LEN) for i in input_pos]
-
-            reference_output = mx.fast.rope(
-                x.transpose(0, 2, 1, 3),
-                dims=HEAD_DIM,
-                traditional=traditional,
-                base=BASE,
-                scale=1.0,
-                offset=input_pos_mx,
-            ).transpose(0, 2, 1, 3)
-            user_output = user_layer(x, input_pos_user)
-            assert_allclose(
-                user_output,
-                reference_output,
-                precision,
-                atol=5e-6 if precision == mx.float32 else 1e-3,
-            )
+@pytest.mark.skipif(
+    not qwen3_0_6b_model_exists(), reason="Qwen3-0.6B-4bit model not found"
+)
+def test_utils_qwen3_0_6b():
+    pass
 
 
-@pytest.mark.parametrize("stream", AVAILABLE_STREAMS, ids=AVAILABLE_STREAMS_IDS)
-@pytest.mark.parametrize("traditional", [False, True], ids=["default", "traditional"])
-@pytest.mark.parametrize("precision", PRECISIONS, ids=PRECISION_IDS)
-def test_task_1_rope_multiple_offsets(
-    stream: mx.Stream, traditional: bool, precision: mx.Dtype
-):
-    rope_helper(stream, traditional, precision)
+@pytest.mark.skipif(not qwen3_4b_model_exists(), reason="Qwen3-4B-4bit model not found")
+def test_utils_qwen3_4b():
+    pass
 
 
-def attention_helper(
-    stream: mx.Stream, H_q, H, L, E, S, BATCH, use_flash_attention: bool = False
-):
-    precision = mx.float32
-    with mx.stream(stream):
-        q_shape = (BATCH, H_q, L, E)
-        kv_shape = (BATCH, H, S, E)
-        scale = 0.8
-        for _ in range(100):
-            query = mx.random.uniform(shape=q_shape, dtype=precision)
-            key = mx.random.uniform(shape=kv_shape, dtype=precision)
-            value = mx.random.uniform(shape=kv_shape, dtype=precision)
-            mask = mx.random.uniform(shape=(BATCH, 1, L, S), dtype=precision)
-
-            reference_output_1 = mx.fast.scaled_dot_product_attention(
-                q=query,
-                k=key,
-                v=value,
-                scale=scale,
-                mask=mask,
-            )
-            reference_output_2 = mx.fast.scaled_dot_product_attention(
-                q=query,
-                k=key,
-                v=value,
-                scale=scale,
-            )
-            if use_flash_attention:
-                user_output_1 = flash_attention(
-                    query,
-                    key,
-                    value,
-                    scale=scale,
-                    mask=mask,
-                )
-                user_output_2 = flash_attention(
-                    query,
-                    key,
-                    value,
-                    scale=scale,
-                )
-            else:
-                user_output_1 = scaled_dot_product_attention_grouped(
-                    query,
-                    key,
-                    value,
-                    scale=scale,
-                    mask=mask,
-                )
-                user_output_2 = scaled_dot_product_attention_grouped(
-                    query,
-                    key,
-                    value,
-                    scale=scale,
-                )
-            mx.eval(user_output_1)
-            mx.eval(user_output_2)
-            assert_allclose(
-                user_output_2,
-                reference_output_2,
-                precision=mx.bfloat16,
-                message="no mask",
-            )
-            assert_allclose(
-                user_output_1,
-                reference_output_1,
-                precision=mx.bfloat16,
-                message="with mask",
-            )
+@pytest.mark.skipif(
+    not qwen3_1_7b_model_exists(), reason="Qwen3-1.7B-4bit model not found"
+)
+def test_utils_qwen3_1_7b():
+    pass
 
 
-def test_task_1_flash_attention_with_mask_cpu_small():
-    attention_helper(mx.cpu, 6, 3, 2, 5, 3, 1, use_flash_attention=True)
-
-
-def test_task_1_flash_attention_with_mask_cpu():
-    attention_helper(mx.cpu, 18, 6, 7, 5, 3, 10, use_flash_attention=True)
-
-
-def test_task_1_flash_attention_with_mask_cpu_large():
-    attention_helper(mx.cpu, 28, 4, 16, 128, 16, 3, use_flash_attention=True)
-
-
-def test_task_1_flash_attention_with_mask_gpu_extra_small():
-    attention_helper(mx.gpu, 1, 1, 5, 7, 4, 1, use_flash_attention=True)
-
-
-def test_task_1_flash_attention_with_mask_gpu_small():
-    attention_helper(mx.gpu, 6, 3, 2, 5, 3, 1, use_flash_attention=True)
-
-
-def test_task_1_flash_attention_with_mask_gpu():
-    attention_helper(mx.gpu, 18, 6, 7, 5, 3, 10, use_flash_attention=True)
-
-
-def test_task_1_flash_attention_with_mask_gpu_large():
-    attention_helper(mx.gpu, 28, 4, 16, 128, 16, 3, use_flash_attention=True)
-
-
-def test_task_1_attention_with_mask_cpu_small():
-    attention_helper(mx.cpu, 6, 3, 2, 5, 3, 1, use_flash_attention=False)
-
-
-def test_task_1_attention_with_mask_cpu():
-    attention_helper(mx.cpu, 18, 6, 7, 5, 3, 10, use_flash_attention=False)
-
-
-def test_task_1_attention_with_mask_cpu_large():
-    attention_helper(mx.cpu, 28, 4, 16, 128, 16, 3, use_flash_attention=False)
-
-
-def test_task_1_attention_with_mask_gpu_extra_small():
-    attention_helper(mx.gpu, 1, 1, 5, 7, 4, 1, use_flash_attention=False)
-
-
-def test_task_1_attention_with_mask_gpu_small():
-    attention_helper(mx.gpu, 6, 3, 2, 5, 3, 1, use_flash_attention=False)
-
-
-def test_task_1_attention_with_mask_gpu():
-    attention_helper(mx.gpu, 18, 6, 7, 5, 3, 10, use_flash_attention=False)
-
-
-def test_task_1_attention_with_mask_gpu_large():
-    attention_helper(mx.gpu, 28, 4, 16, 128, 16, 3, use_flash_attention=False)
-
-
-def test_task_2_batching_kv_cache():
-    cache = BatchingKvCache(max_active_requests=3, max_seq_len=8)
-
-    slot0 = TinyKvFullCache()
-    slot0.update_and_fetch(
-        mx.array([[[[10.0]]]], dtype=mx.float32),
-        mx.array([[[[110.0]]]], dtype=mx.float32),
-    )
-
-    slot2 = TinyKvFullCache()
-    slot2.update_and_fetch(
-        mx.array([[[[20.0], [21.0]]]], dtype=mx.float32),
-        mx.array([[[[120.0], [121.0]]]], dtype=mx.float32),
-    )
-
-    cache.add_request(slot0, 0)
-    cache.add_request(slot2, 2)
-
-    keys = mx.array(
-        [
-            [[[12.0], [13.0]]],
-            [[[0.0], [0.0]]],
-            [[[22.0], [23.0]]],
-        ],
-        dtype=mx.float32,
-    )
-    values = mx.array(
-        [
-            [[[112.0], [113.0]]],
-            [[[0.0], [0.0]]],
-            [[[122.0], [123.0]]],
-        ],
-        dtype=mx.float32,
-    )
-
-    batched_keys, batched_values, seq_len, mask = cache.update_and_fetch(
-        keys, values, mask_length=2
-    )
-
-    expected_keys = mx.array(
-        [
-            [[[0.0], [10.0], [12.0], [13.0]]],
-            [[[0.0], [0.0], [0.0], [0.0]]],
-            [[[20.0], [21.0], [22.0], [23.0]]],
-        ],
-        dtype=mx.float32,
-    )
-    expected_values = mx.array(
-        [
-            [[[0.0], [110.0], [112.0], [113.0]]],
-            [[[0.0], [0.0], [0.0], [0.0]]],
-            [[[120.0], [121.0], [122.0], [123.0]]],
-        ],
-        dtype=mx.float32,
-    )
-    expected_mask = mx.array(
-        [
-            [[[-mx.inf, 0.0, 0.0, -mx.inf], [-mx.inf, 0.0, 0.0, 0.0]]],
-            [
-                [
-                    [-mx.inf, -mx.inf, -mx.inf, -mx.inf],
-                    [-mx.inf, -mx.inf, -mx.inf, -mx.inf],
-                ]
-            ],
-            [[[0.0, 0.0, 0.0, -mx.inf], [0.0, 0.0, 0.0, 0.0]]],
-        ],
-        dtype=mx.float32,
-    ).reshape(3, 1, 2, 4)
-
-    assert seq_len is None
-    assert_allclose(batched_keys, expected_keys, mx.float32)
-    assert_allclose(batched_values, expected_values, mx.float32)
-    assert_allclose(mask, expected_mask, mx.float32)
-
-
-def helper_test_task_3(
-    model_name: str,
-    seq_len: int,
-    iters: int = 1,
-):
-    """Tests for continuous batching of decode requests."""
-    requests = 4
-    max_seq_len = seq_len
-
+def helper_test_task_3(model_name: str, iters: int = 10):
     mlx_model, tokenizer = load(model_name)
     model = Qwen3ModelWeek2(mlx_model)
     for _ in range(iters):
-        cache = [
-            BatchingKvCache(requests, max_seq_len)
-            for _ in range(model.num_hidden_layers)
-        ]
-        # Start each request at a staggered token index.
-        staggered_start = [seq_len * i // requests for i in range(requests)]
-        inputs = mx.random.randint(0, tokenizer.vocab_size, (requests, seq_len))
-        ref_outputs = mlx_model(inputs)
-        for offset in range(seq_len + staggered_start[-1]):
-            seq_idx = [offset - start for start in staggered_start]
-
-            # Requests join at the staggered start, and leave when they reach seq_len.
-            for request_id, sidx in enumerate(seq_idx):
-                if sidx == 0:
-                    for c in cache:
-                        c.add_request(TinyKvFullCache(), request_id)
-                elif sidx == seq_len:
-                    for c in cache:
-                        c.remove_request(request_id)
-
-            next_tokens = []
-            next_offsets = []
-            for request_id, sidx in enumerate(seq_idx):
-                if 0 <= sidx < seq_len:
-                    next_tokens.append(inputs[request_id, sidx].item())
-                    next_offsets.append(sidx)
-                else:
-                    next_tokens.append(0)
-                    next_offsets.append(0)
-
-            user_out = model(
-                inputs=mx.array(next_tokens, dtype=mx.int32).reshape(-1, 1),
-                offset=mx.array(next_offsets, dtype=mx.int32),
-                cache=cache,
-            )
-
-            for request_id, sidx in enumerate(seq_idx):
-                if 0 <= sidx < seq_len:
-                    user_out_r = user_out[request_id, 0, :]
-                    ref_out_r = ref_outputs[request_id, sidx, :]
-                    user_out_r = user_out_r - mx.logsumexp(
-                        user_out_r, keepdims=True
-                    )
-                    ref_out_r = ref_out_r - mx.logsumexp(ref_out_r, keepdims=True)
-                    assert_allclose(
-                        user_out_r,
-                        ref_out_r,
-                        precision=mx.bfloat16,
-                        rtol=0.1,
-                        atol=1.0,
-                    )
+        cache = [TinyKvFullCache() for _ in range(model.num_hidden_layers)]
+        input = mx.random.randint(low=0, high=tokenizer.vocab_size, shape=(1, 10))
+        user_output = model(input, 0, cache)
+        ref_output = mlx_model(input)
+        user_output = user_output - mx.logsumexp(user_output, keepdims=True)
+        ref_output = ref_output - mx.logsumexp(ref_output, keepdims=True)
+        assert_allclose(
+            user_output, ref_output, precision=mx.bfloat16, rtol=0.1, atol=1.0
+        )
 
 
 @pytest.mark.skipif(
     not qwen3_0_6b_model_exists(), reason="Qwen3-0.6B-4bit model not found"
 )
 def test_task_3_qwen3_0_6b():
-    helper_test_task_3("Qwen/Qwen3-0.6B-MLX-4bit", seq_len=3)
+    helper_test_task_3("Qwen/Qwen3-0.6B-MLX-4bit", 5)
+
+
+@pytest.mark.skipif(not qwen3_4b_model_exists(), reason="Qwen3-4B-4bit model not found")
+def test_task_3_qwen3_4b():
+    helper_test_task_3("Qwen/Qwen3-4B-MLX-4bit", 1)
 
 
 @pytest.mark.skipif(
-    not qwen3_4b_model_exists(), reason="Qwen3-4B-4bit model not found"
+    not qwen3_1_7b_model_exists(), reason="Qwen3-1.7B-4bit model not found"
 )
-def test_task_3_qwen3_4b():
-    helper_test_task_3(
+def test_task_3_qwen3_1_7b():
+    helper_test_task_3("Qwen/Qwen3-1.7B-MLX-4bit", 3)
+
+
+def helper_test_task_4(
+    model_name: str,
+    seq_len: int,
+    iters: int = 1,
+):
+    mlx_model, tokenizer = load(model_name)
+    model = Qwen3ModelWeek2(mlx_model)
+    for _ in range(iters):
+        inputs = mx.random.randint(0, tokenizer.vocab_size, (1, seq_len))
+        ref_outputs = mlx_model(inputs)
+        decode_cache = [TinyKvFullCache() for _ in range(model.num_hidden_layers)]
+        for offset in range(seq_len):
+            user_out = model(
+                inputs=inputs[:, offset : offset + 1],
+                offset=offset,
+                cache=decode_cache,
+            )
+            ref_out = ref_outputs[:, offset : offset + 1, :]
+            user_out = user_out - mx.logsumexp(user_out, keepdims=True)
+            ref_out = ref_out - mx.logsumexp(ref_out, keepdims=True)
+            assert_allclose(
+                user_out, ref_out, precision=mx.bfloat16, rtol=0.1, atol=1.0
+            )
+
+
+@pytest.mark.skipif(
+    not qwen3_0_6b_model_exists(), reason="Qwen3-0.6B-4bit model not found"
+)
+def test_task_4_qwen3_0_6b():
+    helper_test_task_4("Qwen/Qwen3-0.6B-MLX-4bit", seq_len=3)
+
+
+@pytest.mark.skipif(not qwen3_4b_model_exists(), reason="Qwen3-4B-4bit model not found")
+def test_task_4_qwen3_4b():
+    helper_test_task_4(
         "Qwen/Qwen3-4B-MLX-4bit",
         seq_len=3,
     )
@@ -338,8 +109,29 @@ def test_task_3_qwen3_4b():
 @pytest.mark.skipif(
     not qwen3_1_7b_model_exists(), reason="Qwen3-1.7B-4bit model not found"
 )
-def test_task_3_qwen3_1_7b():
-    helper_test_task_3(
-        "Qwen/Qwen3-1.7B-MLX-4bit",
-        seq_len=3,
-    )
+def test_task_4_qwen3_1_7b():
+    helper_test_task_4("Qwen/Qwen3-1.7B-MLX-4bit", seq_len=3)
+
+
+class FakeEmbedding:
+    def __call__(self, inputs):
+        return mx.stack([inputs, inputs + 1], axis=-1).astype(mx.float32)
+
+    def as_linear(self, hidden):
+        return hidden
+
+
+@pytest.mark.parametrize("model_class", [Qwen3ModelWeek2, Qwen3ModelWeek3])
+@pytest.mark.parametrize("logits_to_keep,expected_length", [(1, 1), (None, 4)])
+def test_logits_to_keep_controls_output_length(
+    model_class, logits_to_keep, expected_length
+):
+    model = model_class.__new__(model_class)
+    model.num_hidden_layers = 0
+    model.embedding = FakeEmbedding()
+    model.layers_inner = []
+    model.norm = lambda hidden: hidden
+    model.w_lm_head = None
+    inputs = mx.array([[1, 2, 3, 4]])
+    result = model(inputs, 0, [], logits_to_keep=logits_to_keep)
+    assert result.shape == (1, expected_length, 2)
